Tower Placement Complete
The general idea for allowing placement of towers is now up. There is also a test tower for it to be tested with. That scene also contains the "idea" to be used behind each tower.

Some issues with towers however that I can see is that varying tower sizes means it will be hard to deny tower overlap unless we standardize size.