Tower Placeable Indicator + Code Trim
There is now a functioning indicator for if a tower can be placed or not. Red if you can't, green if you can. Seems well functioning and a better representation of the space a tower takes up with their areadenial.

I also trimmed some redundant code I had for position setting.