Fixed Tower Attack and Function being called excessively
Fixed issue where towers would not detect anything entering its range.

The problem occured due to the units making use of a StaticBody3D which does not collide when moved. It was replaced with a CharacterBody3D which collides while moving.

From what can be seen, it was working prior only in preview as our position changing created instances where the unit was considered static thus allowing it to enter the body and be detected.

I also added a variable called only_once that is used to prevent addcollision and removecollision being run excessively. If you have a better idea, please change this!