Tower upgrade paths added as well as upgrade status indicator.
I fixed the mouse not being captured and added the handler for upgrade paths for towers as well as documentation.

I also added to the tower model to now include an upgrade indicator that only appears when your mouse is unlocked and hovering on a specific tower. Currently green means you can DOUBLE CLICK to upgrade, red you don't have enough money to upgrade, and gold its fully upgraded.

I also added glow to the world environment to allow the upgrade indicators to glow. Doesn't matter either way to me if we keep it or not as  it wasn't making the indicator as glowy as I wished.