Added Constraints to Camera
Users are now trapped in the box of the map. They can still fly into the bases and etc, but they cant clip into the ground.

Thinking we should just let them and maybe put just a fun thing in the bases if they clip in. Make it seem like we meant to allow them.