Added Game Over Handler and Health Bars
There is now a game over function within Global that will handle things once it is called for a game over. It can be provided an array created as described to store data that is not a default. Only type implemented right now is score.

With game_over() now in place, any changes to Currency variables should be also changed in game_over() in Global. If you have an idea on how to make this not necessary, by all means.

Some currency generation was also changed. The timing for tC gen and grugarian gen are now different and can be modified seperately. I also made it a little longer to gen.

Finally, units now have healthbars. They don't exactly lower but they swap between coloring to indicate state of health of each unit. Currently goes off of thirds of health. More just added it to show concept and have it there. Future units will either have that health bar as well or display health through changes in coloring/texture in model. Really depends on if I get a good idea and have the time. Else, woe health bar upon ye in some fashion.