Added Tower and Unit Placeholders and Changed preview_placement() args.
There are now placeholders for the tower and unit archetypes I will be adding. I also moved all towers to be in one location like with units. You should be able to understand what they will do by their name.

I also changed the preview_placement() function. The UI should now preload the models for the towers and provide the variable that holds the preload to preview_placment(). Better to preload than load again and again.